Atmospheric audio clips
Free for use from http://soundbible.com/.  Where audio clips have a first and last name, credit must be given to the audio creator.  For files where, "public domain" and "personal use" appear in the file name, no credit is required.  To use "personal use" files, we can't charge for the use of the game.